requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
# langchain-embeddings
langchain-core
langchain-community
chromadb
beautifulsoup4
lxml
requests
python-dotenv
openai
tiktoken
pandas
crawl4ai
playwright